delete helpers.go
